Add new tip for diagnosing problems
I fired up my first application on Fly.io and it took me a while to figure out why my assets were not building as I expected. It was a simply an issue of not having an `npm install` step in my Dockerfile. I thought this tip may ease some frustration for someone down the line.

I didn't include any details about how one might include `nodejs` or the like but maybe that is warranted, didn't want to be too prescriptive here.